HBASE-11229 Change block cache percentage metrics to be doubles rather than ints
In HBASE-8370 "Report data block cache hit rates apart from aggregate cache hit rates"
discussion highlights small changes in caching percentage can make for large
changes in cache contents; e.g. lots of data blocks may be getting evicited
but because meta blocks -- index and blooms -- are staying in cache with
high hit rates, a large data block eviction may show as a small change in
overall cache percentage (e.g. from 99.99 to 99.11).

Changes the getBlockCacheHitPercent from an int to a double.
Also changes name of the jmx metric blockCountHitPercent to be
blockCacheCountHitPercent instead.

May be incompatible change.

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/MetricsRegionServerWrapperImpl.java
@@ -242,7 +242,7 @@ class MetricsRegionServerWrapperImpl
   }
 
   @Override
-  public int getBlockCacheHitPercent() {
+  public double getBlockCacheHitPercent() {
     if (this.cacheStats == null) {
       return 0;
     }
